move common code to common place
