+++ check_manager/src/api_interface.py
@@ -1,7 +1,8 @@
-from typing import Any, Final, Iterable
+from typing import Final, Iterable
 from urllib import parse
 from fastapi import Request
 
+from api_utils.exceptions import CheckInternalError
 from check_backends.check_backend import (
     CheckIdError,
     CheckIdNonUniqueError,
@@ -9,12 +10,10 @@ from check_backends.check_backend import (
 )
 from exceptions import (
     CheckException,
-    CheckExceptions,
-    CheckInternalError,
     CheckConnectionError,
     JsonValidationError,
 )
-from api_utils.json_api_types import APIErrorResponse, Error
+from api_utils.json_api_types import Error
 
 
 ERROR_CODE_KEY: Final[str] = "error"
@@ -32,34 +31,7 @@ REMOVE_CHECK_PATH: Final[str] = ROUTE_PREFIX + "/checks/{check_id}"
 GET_CHECKS_PATH: Final[str] = ROUTE_PREFIX + "/checks/"
 
 
-def get_status_code_and_errors(exception: Exception) -> tuple[int, list[Error]]:
-    match exception:
-        case CheckExceptions():
-            exceptions = exception
-        case CheckException():
-            exceptions = CheckExceptions(
-                status=int(exception.error.status), exceptions=[exception]
-            )
-        case _:
-            exceptions = CheckExceptions(
-                status=500,
-                exceptions=[CheckInternalError.create("Internal server error")],
-            )
-    return exceptions.status, [exc.error for exc in exceptions.exceptions]
-
-
-def get_exceptions(
-    status_code: int, content: dict[str, Any]
-) -> CheckException | CheckExceptions:
-    error_response = APIErrorResponse.model_validate(content)
-    exceptions = [_get_exception(error) for error in error_response.errors]
-    if len(exceptions) == 1:
-        return exceptions[0]
-    else:
-        return CheckExceptions(status=status_code, exceptions=exceptions)
-
-
-def _get_exception(error: Error) -> CheckException:
+def get_exception(error: Error) -> CheckException:
     match error.code:
         case JsonValidationError.__name__:
             return JsonValidationError(error)
